+++ frontend/.env
@@ -1,0 +1,9 @@
+# .env
+FB_API_KEY=AIzaSyBfkajd8fUdQ5NeqiBM1p3WVbOD8pgrOX8
+FB_AUTH_DOMAIN=adventure-board-40e02.firebaseapp.com
+FB_PROJECT_ID=adventure-board-40e02
+FB_STORAGE_BUCKET=adventure-board-40e02.firebasestorage.app
+FB_MESSAGING_SENDER_ID=1065719539107
+FB_APP_ID=1:1065719539107:web:bb0adcf838091c104d2616
+FB_MEASUREMENT_ID=G-ZKJ4BGQ9FX
+FB_VAPID_KEY=BHdufWLs6iJn8hQ_ZtNSRtEpF9mVdXjlNhvoqWx2fOkWe-FpDNX0T2_bPr3eeV-oTfpz-0MLIKdbiatQ4nEF6xg
